requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
imutils==0.5.3
joblib==0.15.1
kiwisolver==1.3.1
matplotlib==3.4.3
networkx==2.5
numpy==1.24.4
opencv-python==4.8.1.78
packaging==25.0
Pillow==8.3.2
psycopg2-binary==2.9.10
pyparsing==2.4.7
pytesseract==0.3.13
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.19.3
scikit-learn==0.24.2
scipy==1.10.1
six==1.15.0
threadpoolctl==2.0.0
tifffile==2020.9.3